Created Policy Editor. Added updateAll block to backend.
Started on Policy Editor Page.
Created dictionary objects for the forms that have a related table.

Changed related table's column property to be its id.
Before: Adding role to user, we added the role name
Now: We add the role's id to user, and store the role information in a object to display it in html